[병합] Member 관련 Branch 임시로 합치기
- 10-member-BE-jeong
- 10-member-BE-heesom

commit 44f7040e74ba378eaa5353c56861c1e170dea65c
Author: 진희솜 <heesom.hs@gmail.com>
Date:   Tue Jul 25 11:06:42 2023 +0900

    [테스트] access 재발급에 관한 테스트케이스 작성

    - refresh_값이_empty일때_체크
    - refresh_값이_유효하지_않을때_체크
    - refresh_유효기간이_지났을때_체크

commit 2240324145fa2a7c677f6a4c9bde8ec44d12d130
Author: 진희솜 <heesom.hs@gmail.com>
Date:   Tue Jul 25 10:46:38 2023 +0900

    [기능] refresh 토큰을 받고 access 토큰 반환하는 기능

commit 02fd856beb3112741ed25fe3131abb1a849739ea
Author: Heesom <heesom.hs@gmail.com>
Date:   Mon Jul 24 20:43:12 2023 +0900

    [테스트] CustomJWTException에 대한 테스트 케이스 작성

    - 해킹 위험이 있는 토큰에 관한 예외 발생 테스트 케이스 작성
    - 유효 기간이 지난 토큰에 관한 예외 발생 테스트 케이스 작성
    - 시간 경과 후 다른 토큰이 생성되는지 확인

commit 95746f5bfcb12131254f38e16d2f45de75524d57
Author: Heesom <heesom.hs@gmail.com>
Date:   Mon Jul 24 20:40:53 2023 +0900

    [버그] jwt에서 유효 시간이 경과된 exception을 잡지 못하는 에러 해결

    - 유효 기간이 지난 토큰과 잘못 만들어진 토큰의 예외 처리를 분리

commit bf878e0415e1d64d8f477a7b02b6e057561e8211
Author: Heesom <heesom.hs@gmail.com>
Date:   Mon Jul 24 18:06:51 2023 +0900

    [기능] 회원 탈퇴

    - mysql에 대한 회원 삭제만 진행하기로 함
    - 이미 탈퇴한 회원에 대한 CustomDBException 생성

commit 418494266eaf6653c8135c1436170cb4912ad2fc
Merge: 3efbb88 d161469
Author: 김민수 <mskim.spring@gmail.com>
Date:   Sat Jul 22 17:49:01 2023 +0900

    Merge branch '10-member-BE-jeong' into '10-member-BE'

    [ 10-member-BE-jeong -> 10-member-BE ]

    See merge request s09-webmobile1-sub2/S09P12D109!7

.

+++ backend/src/main/java/com/codragon/sclive/service/UserServiceImpl.java
@@ -1,6 +1,9 @@
 package com.codragon.sclive.service;
 
 import com.codragon.sclive.dao.UserDao;
+import com.codragon.sclive.dto.UserReqDto;
+import com.codragon.sclive.exception.CustomDBException;
+import com.codragon.sclive.exception.DBErrorCode;
 import com.codragon.sclive.jwt.JWTUtil;
 import com.codragon.sclive.jwt.Jwt;
 import com.codragon.sclive.mapper.UserMapper;
@@ -11,6 +14,8 @@ import org.springframework.stereotype.Service;
 
 import javax.servlet.http.Cookie;
 import javax.servlet.http.HttpServletResponse;
+
+
 @Slf4j
 @Service
 public class UserServiceImpl implements UserService {
@@ -59,12 +64,16 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public int deleteUser(String accessToken) {
+    public void deleteUser(String accessToken) {
         String email = jwt.getEmailFromToken(accessToken);
-        int isDeleted = jwtUtil.deleteUserRefreshToken(email);
+        jwtUtil.deleteUserRefreshToken(email);
+
+        // 이미 탈퇴한 회원이라면
+        UserDao mysqlInfo = this.getUserInfo(email);
+        if (mysqlInfo == null) {
+            throw new CustomDBException(DBErrorCode.ALREADY_DELETED);
+        }
         userMapper.deleteUser(email);
-        return isDeleted;
-        // TODO:redis에러, mysql 삭제 트랜잭션 처리
     }
 
     @Override
@@ -81,7 +90,7 @@ public class UserServiceImpl implements UserService {
         // 입력된 비번과 디비의 암호화된 비번이 같은지 확인.
         boolean passwordMatched = passwordEncoder.matches(password, userDao.getPassword());
 
-        if(passwordMatched){ //유효한 패스워드이다.
+        if (passwordMatched) { //유효한 패스워드이다.
             //토큰 발급
             String accessToken = jwt.createAccessToken(userDao.getEmail(), userDao.getNickname());
             String refreshToken = jwt.createRefreshToken(userDao.getEmail(), userDao.getNickname());
@@ -95,7 +104,7 @@ public class UserServiceImpl implements UserService {
             //redis에 RefreshToken 저장
             jwtUtil.saveUserRefreshToken(userDao.getEmail(), refreshToken);
             return ResponseEntity.status(200).body("Success");
-        } else{
+        } else {
             return ResponseEntity.status(200).body("Fail");
         }
     }
